pyproject.toml
[project]
name = "hw5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "gymnasium[box2d]>=1.2.2",
    "langchain>=1.0.8",
    "langchain-core>=1.1.0",
    "litellm>=1.80.5",
    "mlflow>=3.6.0",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "optunahub>=0.4.0",
    "stable-baselines3[extra]>=2.7.0",
]
